+++ benchmarking/example_model_upload.py
@@ -1,7 +1,7 @@
 import linktransformer as lt
 
 
-model_path="/mnt/122a7683-fa4b-45dd-9f13-b18cc4f4a187/deeprecordlinkage/linktransformer/models/test_model/0"
+model_path=""
 
 ###Load the model
 model=lt.load_model(model_path)