+++ src/linktransformer/infer.py
@@ -99,7 +99,8 @@ def merge(
         strings_right = df2[right_on].tolist()
 
     ## Load the model
-    model = load_model(model)
+    if openai_key is None:
+        model = load_model(model)
 
     ## Infer embeddings for df1
     embeddings1 = infer_embeddings(strings_left, model, batch_size=batch_size, openai_key=openai_key)
@@ -488,5 +489,168 @@ def all_pair_combos_evaluate(df,model,left_on,right_on,openai_key=None):
     df_concat = pd.DataFrame({"left_on":left_on_all_pairs,"right_on":right_on_all_pairs,"score":cosine_similarity_12_all_pairs})
 
     return df_concat
+
+
+
+def merge_knn(
+    df1: DataFrame,
+    df2: DataFrame,
+    merge_type: str = '1:1',
+    on: Optional[Union[str, List[str]]] = None,
+    model: str = "all-MiniLM-L6-v2",
+    left_on: Optional[Union[str, List[str]]] = None,
+    right_on: Optional[Union[str, List[str]]] = None,
+    k: int = 1,
+    suffixes: Tuple[str, str] = ('_x', '_y'),
+    use_gpu: bool = False,
+    batch_size: int = 128,
+    openai_key: Optional[str] = None,
+    drop_sim_threshold: float = None
+) -> DataFrame:
+    """
+    Merge two dataframes using language model embeddings. This function would support k nearest neighbors matching for each row in df1.
+    Merge is a special case of this function when k=1.
+    :param df1 (DataFrame): First dataframe (left).
+    :param df2 (DataFrame): Second dataframe (right).
+    :param on (Union[str, List[str]], optional): Column(s) to join on in df1. Defaults to None.
+    :param model (str): Language model to use.
+    :param left_on (Union[str, List[str]], optional): Column(s) to join on in df1. Defaults to None.
+    :param right_on (Union[str, List[str]], optional): Column(s) to join on in df2. Defaults to None.
+    :param k (int): Number of nearest neighbors to match for each row in df1. Defaults to 1.
+    :param suffixes (Tuple[str, str]): Suffixes to use for overlapping columns. Defaults to ('_x', '_y').
+    :param use_gpu (bool): Whether to use GPU. Not supported yet. Defaults to False.
+    :param batch_size (int): Batch size for inferencing embeddings. Defaults to 128.
+    :param openai_key (str, optional): OpenAI API key for InferKit API. Defaults to None.
+    :return: DataFrame: The merged dataframe.
+    """
+
+    ## Set common columns as on if not specified
+    if on is None:
+        on = list(set(df1.columns).intersection(set(df2.columns)))
+
+    ## If left_on or right_on is not specified, set it to on
+    if left_on is None:
+        left_on = on
+    if right_on is None:
+        right_on = on
+
+    on = None
+
+    ### If how is 1:m, then we want to merge df1 to df2
+    ### Check if how is valid
+    if merge_type not in ["1:m", "m:1", "1:1"]:
+        raise ValueError(f"Invalid merge type: {merge_type}")
+
+    if merge_type == "1:m":
+        if df1[left_on].duplicated().any():
+           print("Warning: Keys in df1 are not unique")
+
+    if merge_type == "m:1":
+        ## Check if keys in df2 are unique
+        if df2[right_on].duplicated().any():
+            print("Warning: Keys in df2 are not unique")
+    if merge_type == "1:1":
+        ## Check if keys in df1 are unique
+        if df1[left_on].duplicated().any():
+           print("Warning: Keys in df1 are not unique")
+        ## Check if keys in df2 are unique
+        if df2[right_on].duplicated().any():
+           print("Warning: Keys in df1 are not unique")
+
+    df1 = df1.copy()
+    df2 = df2.copy()
+    ## give ids to each df
+    ##Ensure that there is no id_lt column in df1 or df2
+    if "id_lt" in df1.columns:
+        raise ValueError(f"Column id_lt already exists in df1, please rename it to proceed")
+    if "id_lt" in df2.columns:
+        raise ValueError(f"Column id_lt already exists in df2,please rename it to proceed")
+
+    df1.loc[:, "id_lt"] = np.arange(len(df1))
+    df2.loc[:, "id_lt"] = np.arange(len(df2))
+
+    if isinstance(right_on, list):
+        strings_right = serialize_columns(df2, right_on, model=model)
+    if isinstance(left_on, list):
+        strings_left = serialize_columns(df1, left_on, model=model)
+    else:
+        strings_left = df1[left_on].tolist()
+        strings_right = df2[right_on].tolist()
+    
+    ## Load the model
+    model = load_model(model)
+
+    ## Infer embeddings for df1
+    embeddings1 = infer_embeddings(strings_left, model, batch_size=batch_size, openai_key=openai_key)
+    ## Infer embeddings for df2
+    embeddings2 = infer_embeddings(strings_right, model, batch_size=batch_size, openai_key=openai_key)
+
+    ### Expand dim if embeddings are 1d (numpy)
+    if len(embeddings1.shape) == 1:
+        embeddings1 = np.expand_dims(embeddings1, axis=0)
+    if len(embeddings2.shape) == 1:
+        embeddings2 = np.expand_dims(embeddings2, axis=0)
+
+    ## Normalize embedding tensors using numpy
+
+    embeddings1 = embeddings1 / np.linalg.norm(embeddings1, axis=1, keepdims=True)
+    embeddings2 = embeddings2 / np.linalg.norm(embeddings2, axis=1, keepdims=True)
+
+    ## Create index
+    if use_gpu:
+        raise ValueError(f"GPU not supported yet")
+    else:
+        index = faiss.IndexFlatIP(embeddings1.shape[1])
+
+    ## Add to index depending on merge type
+    if merge_type == "1:m":
+        index.add(embeddings2)
+    elif merge_type == "m:1":
+        index.add(embeddings2)
+    elif merge_type == "1:1":
+        index.add(embeddings2)
+
+    ## Search index
+    if merge_type == "1:m":
+        D, I = index.search(embeddings1, k)
+    elif merge_type == "m:1":
+        D, I = index.search(embeddings1, k)
+    elif merge_type == "1:1":
+        D, I = index.search(embeddings1, k)
+
+    ## Check nearest neighbor of the first text in df1 as a test
+    df1 = df1.reset_index(drop=True)
+    df2 = df2.reset_index(drop=True)
+
+    ## Fuzzily merge the dfs based on the faiss index queries
+    ###Each I sublst is a list of k nearest neighbors for each row in df1 in terms of indices of df2
+    ###We need to expand the rows of df1 and df2 to match the number of rows in df1
+    ###We also need to expand the scores to match the number of rows in df1
+
+    ###First, expand the rows of df1
+    df1_expanded = df1.loc[np.repeat(df1.index.values, k)].reset_index(drop=True)
+
+    ###Now, expand the rows of df2
+    df2_expanded = df2.iloc[I.flatten()].reset_index(drop=True)
+
+
+    ###Now, merge the expanded dfs
+    df_lm_matched = df1_expanded.merge(df2_expanded, left_index=True, right_index=True, how="inner")
+
+    ### Add score column
+    df_lm_matched["score"] =  D.flatten()
+
+    print(f"LM matched on key columns - left: {left_on}{suffixes[0]}, right: {right_on}{suffixes[1]}")
+
+    if drop_sim_threshold is not None:
+        df_lm_matched = df_lm_matched[df_lm_matched["score"]>=drop_sim_threshold]
+        print(f"Dropped rows with similarity below {drop_sim_threshold}")
+
+    return df_lm_matched
+
+
+
+
+